+++ reduct_cli/__main__.py
@@ -1,5 +1,5 @@
 """Entry point"""
-from reduct_cli.main import alias
+from reduct_cli import cli
 
 if __name__ == "__main__":
-    alias()
+    cli(obj={})
