+++ reduct_cli/consoles.py
@@ -1,0 +1,4 @@
+from rich.console import Console
+
+error_console = Console(stderr=True, style="bold red")
+console = Console()
