--- reduct_cli/main.py
@@ -1,6 +1,0 @@
-"""Main module"""
-
-import click
-from alias import alias
-
-cli = click.CommandCollection(sources=[alias(obj={})])
